rpi_master_service/requirements.txt
spidev>=3.5
RPi.GPIO>=0.7.0
matplotlib>=3.5.0
pandas>=1.4.0
numpy>=1.21.0